+++ TagCloud/Creators/Tag.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagCloud.Layouters
+namespace TagCloud.Creators
 {
     public class Tag
     {
