+++ TagCloud/Layouters/CircularCloudLayouter.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using TagCloud.Creators;
 
 namespace TagCloud.Layouters
 {
